Move exception translation routine out of libpydynd.
