+++ dynd/include/kernels/apply_jit_kernel.hpp
@@ -6,7 +6,6 @@
 #include <utility_functions.hpp>
 #include <arrfunc_from_pyfunc.hpp>
 #include <type_functions.hpp>
-#include <exception_translation.hpp>
 #include <array_assign_from_py.hpp>
 
 #include "../arrfunc_functions.hpp"
